add test method getuseroutcomes

+++ src/test/java/com/fm/internal/services/OutcomeTypeServiceTest.java
@@ -30,8 +30,10 @@ public class OutcomeTypeServiceTest extends AbstractTestNGSpringContextTests {
     @Autowired
     private OutcomeService outcomeService;
 
-    private final static String USER_EMAIL = "user@email";
-    private final static String USER_PASSWORD = "password";
+    private final static String FIRST_USER_EMAIL = "user@email";
+    private final static String FIRST_USER_PASSWORD = "password";
+    private final static String SECOND_USER_EMAIL = "user2@email";
+    private final static String SECOND_USER_PASSWORD = "password";
     private final static String FIRST_ACCOUNT_NAME = "Visa";
     private final static String SECOND_ACCOUNT_NAME = "Mastercard";
     private final static String THIRD_ACCOUNT_NAME = "Мир";
@@ -39,12 +41,15 @@ public class OutcomeTypeServiceTest extends AbstractTestNGSpringContextTests {
 
     @BeforeMethod
     public void setUp() throws Exception {
-        User user = new User(USER_EMAIL, USER_PASSWORD, new UserInfo("name", "surname"));
+        User user = new User(FIRST_USER_EMAIL, FIRST_USER_PASSWORD, new UserInfo("name", "surname"));
+        User secondUser = new User(SECOND_USER_EMAIL, SECOND_USER_PASSWORD, new UserInfo("name", "surname"));
         userService.createUser(user);
+        userService.createUser(secondUser);
         Account[] accounts = {
                 new Account(FIRST_ACCOUNT_NAME, new BigDecimal(29999), null, user),
                 new Account(SECOND_ACCOUNT_NAME, new BigDecimal(3000), null, user),
-                new Account(THIRD_ACCOUNT_NAME, new BigDecimal(5752), null, user)
+                new Account(THIRD_ACCOUNT_NAME, new BigDecimal(5752), null, user),
+                new Account(THIRD_ACCOUNT_NAME, new BigDecimal(1111), null, secondUser)
         };
         for (Account account : accounts) {
             accountService.createAccount(account);
@@ -53,21 +58,24 @@ public class OutcomeTypeServiceTest extends AbstractTestNGSpringContextTests {
         OutcomeType[] types = {
                 new OutcomeType("Еда вне дома", new BigDecimal(3000), user),
                 new OutcomeType("ФФСБ", new BigDecimal(2323), user),
-                new OutcomeType("dfsfsd", new  BigDecimal(3242), user)
+                new OutcomeType("dfsfsd", new  BigDecimal(3242), user),
+                new OutcomeType("Еда вне дома", new BigDecimal(3000), secondUser),
         };
         for (OutcomeType type : types) {
             outcomeTypeService.addOutcomeType(type);
         }
 
         Outcome[] outcomes = {
-                new Outcome(new BigDecimal(2122), LocalDate.now(), LocalTime.now(), accounts[0], types[0]),
-                new Outcome(new BigDecimal(332), LocalDate.now(), LocalTime.now(), accounts[0], types[0]),
-                new Outcome(new BigDecimal(4144), LocalDate.now(), LocalTime.now(), accounts[0], types[1]),
-                new Outcome(new BigDecimal(9999), LocalDate.now(), LocalTime.now(), accounts[0], types[1]),
-                new Outcome(new BigDecimal(50), LocalDate.now(), LocalTime.now(), accounts[0], types[2]),
-                new Outcome(new BigDecimal(1234), LocalDate.now(), LocalTime.now(), accounts[0], types[2]),
-                new Outcome(new BigDecimal(2222), LocalDate.now(), LocalTime.now(), accounts[0], types[2]),
-                new Outcome(new BigDecimal(3999), LocalDate.now(), LocalTime.now(), accounts[0], types[0]),
+                new Outcome(new BigDecimal(1), LocalDate.now(), LocalTime.now(), accounts[0], types[0]),
+                new Outcome(new BigDecimal(2), LocalDate.now(), LocalTime.now(), accounts[0], types[0]),
+                new Outcome(new BigDecimal(3), LocalDate.now(), LocalTime.now(), accounts[0], types[2]),
+                new Outcome(new BigDecimal(4), LocalDate.now(), LocalTime.now(), accounts[0], types[2]),
+                new Outcome(new BigDecimal(5), LocalDate.now(), LocalTime.now(), accounts[0], types[2]),
+                new Outcome(new BigDecimal(6), LocalDate.of(2017, 04, 20), LocalTime.now(), accounts[0], types[0]),
+                new Outcome(new BigDecimal(7), LocalDate.of(2017, 04, 20), LocalTime.now(), accounts[0], types[1]),
+                new Outcome(new BigDecimal(8), LocalDate.of(2017, 04, 20), LocalTime.now(), accounts[0], types[1]),
+                new Outcome(new BigDecimal(9), LocalDate.now(), LocalTime.now(), accounts[3], types[3]),
+                new Outcome(new BigDecimal(10), LocalDate.now(), LocalTime.now(), accounts[3], types[3])
         };
         for (Outcome outcome : outcomes) {
             outcomeService.addOutcome(outcome);
@@ -76,7 +84,7 @@ public class OutcomeTypeServiceTest extends AbstractTestNGSpringContextTests {
 
     @AfterMethod
     public void tearDown() throws Exception {
-        User user = userService.findByEmail(USER_EMAIL);
+        User user = userService.findByEmail(FIRST_USER_EMAIL);
         if (user != null) {
             List<Account> accounts = accountService.findAllUserAccounts(user);
             accounts.stream().forEach(
@@ -86,24 +94,44 @@ public class OutcomeTypeServiceTest extends AbstractTestNGSpringContextTests {
                     }
             );
             outcomeTypeService.getAvailableOutcomeTypes(user).stream().forEach(outcomeType -> outcomeTypeService.deleteOutcomeType(outcomeType));
-            user = userService.findByEmail(USER_EMAIL);
+            user = userService.findByEmail(FIRST_USER_EMAIL);
+            userService.deleteUser(user);
+        }
+        user = userService.findByEmail(SECOND_USER_EMAIL);
+        if (user != null) {
+            List<Account> accounts = accountService.findAllUserAccounts(user);
+            accounts.stream().forEach(
+                    account -> {outcomeService.findAllOutcomesInAccount(account).stream().forEach(
+                        outcome -> outcomeService.deleteOutcome(outcome));
+                        accountService.deleteAccount(account);
+                    }
+            );
+            outcomeTypeService.getAvailableOutcomeTypes(user).stream().forEach(outcomeType -> outcomeTypeService.deleteOutcomeType(outcomeType));
+            user = userService.findByEmail(SECOND_USER_EMAIL);
             userService.deleteUser(user);
         }
     }
 
     @Test
     public void testGetOutcomeTypesValue() throws Exception{
-        User user = userService.findByEmail(USER_EMAIL);
+        User user = userService.findByEmail(FIRST_USER_EMAIL);
         Account account = accountService.findUserAccountByName(user, FIRST_ACCOUNT_NAME);
         Assert.assertNotNull(outcomeTypeService.countOutcomeTypesValueByDate(account, LocalDate.of(2017, 4, 23), LocalDate.of(2017, 4, 26)));
-        outcomeTypeService.countOutcomeTypesValueByDate(account, LocalDate.of(2017, 4, 23), LocalDate.of(2017, 4, 26)).entrySet().stream().forEach(e -> System.out.printf(e.getKey()+" %f%n", e.getValue()));
+        outcomeTypeService.countOutcomeTypesValueByDate(account, LocalDate.of(2017, 4, 20), LocalDate.now()).entrySet().stream().forEach(e -> System.out.printf(e.getKey()+" %f%n", e.getValue()));
     }
 
     @Test
     public void testGetAvailableOutcomeTypes() throws Exception {
-        User user = userService.findByEmail(USER_EMAIL);
+        User user = userService.findByEmail(FIRST_USER_EMAIL);
         List<OutcomeType> availableOutcomeTypes = outcomeTypeService.getAvailableOutcomeTypes(user);
         Assert.assertEquals(availableOutcomeTypes.size(), 9);
     }
 
+    @Test
+    public void testGetUserOutcomes() throws Exception {
+        User user = userService.findByEmail(FIRST_USER_EMAIL);
+        List<Outcome> userOutcomes = outcomeService.getUserOutcomes(user);
+        Assert.assertNotNull(userOutcomes);
+        userOutcomes.stream().forEach(outcome -> System.out.println(outcome.getAmount()));
+    }
 }
